feat: event source stream
